[Bug 2962] truncation of size_t/ptrdiff_t on 64bit targets  - fixed several other warnings (cast-alignment, missing const, missing prorotypes)  - promote use of 'size_t' for values that express a size  - use ptr-to-const for read-only arguments  - make sure SOCKET values are not truncated (win32-specific)  - format string fixes

+++ lib/isc/sockaddr.c
@@ -153,7 +153,7 @@ isc_sockaddr_totext(const isc_sockaddr_t *sockaddr, isc_buffer_t *target) {
 		return (ISC_R_FAILURE);
 	}
 
-	plen = strlen(pbuf);
+	plen = (unsigned int)strlen(pbuf);
 	INSIST(plen < sizeof(pbuf));
 
 	isc_netaddr_fromsockaddr(&netaddr, sockaddr);
